+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes:
 *   Fix AdaBoostClassifier structural test ([#318](https://github.com/AI-SDC/SACRO-ML/pull/318))
 *   Fix user story examples ([#320](https://github.com/AI-SDC/SACRO-ML/pull/320), [#321](https://github.com/AI-SDC/SACRO-ML/pull/321))
 *   Update acro dependency to v0.4.8 ([#322](https://github.com/AI-SDC/SACRO-ML/pull/322))
+*   Add Target loading of probas from persistent storage ([#323](https://github.com/AI-SDC/SACRO-ML/pull/323))
 
 ## Version 1.2.1 (Jul 29, 2024)